Caching & runes
Uses Svelte 5 runes to cache the contests and contest objects in state.svelte.ts.
Updates all pages to only load contest data if it hasn't been loaded already.
Improves performance, reduces load on server, and cleans up some of the hardcoding.

The only object not cached is the scoreboard. Will need to cache it and implement
a scheduled or event-driven refresh. Submissions are also cached, and will need to
be updated as well.