Update team submission list to table
The team detail page had a list of submission times with no details. This changes
it to include a table of submissions instead, with columns for time, problem,
language and judgement. It's basic for now, but much more useful than the prior
list.

Also moves the country name calculation to server-side - but disables it for now
as I'm having trouble with it.