MPEG-TS
Adds support for MPEG-TS using mpegts.js. I thought it could work as a plugin to
video.js but don't have that working yet, so for now it is at the expense of HLS.

Video.js code is just commented out, later we'll look at the stream and switch
player or configuration based on which is available.