Rename contest objects
I initially named all the contest API types as 'xJSON' because they were JSON types
from the Contest API. In reality though it's unimportant what format they are
typically loaded from, they are just regular typescript objects and the naming
pattern is verbose and annoying.

Before we go too much farther, this just renames all the types to remove the 'JSON'.
The Contest object had to be renamed to ContestAPI to avoid overlap. The result is
slightly more readable and shorter code.