Substitute missing images
If a contest doesn't have a banner or logos or photos, they're just omitted from the UI.
The app is reactive to what is available, no problem.

However, if a contest says it has any of these images but then downloading the image
fails (i.e. bad file reference), the UI is littered with little error icons. This just
substitutes any missing images with the ICPC logo for now. We could just fancy and use
a different image, or change based on whether it's a logo or photo etc, but for now at
least the UI still looks clean.