Support for file reference tags
Adds support for file reference tags as per the draft spec, and exposes tags through
the base Image, Banner, and Logo components.

The contest logo at the top left and the team logo underneath are the only two places
right now where we always have a dark background, so I used the tag in these places.
At some point we'll need to get fancier and support light/dark mode in the app, but
this was used to prove that the API and core support is working for now.