Video can only play a single video
The Video component was created with ref?:FileReference[] because that was expedient
at the time, and it would always play the first reference. However, it is only meant
to play a single video, and FileReference[] could contain more than one. (e.g. for
WF Moscow it was going to be one video per team member, and reaction videos are 1
for desktop and 1 for webcam).

This just changes it to ref?:FileReference and the client is responsible for picking
which ref to play.