Reactive team and problem pages
The only connection to the backend contest API server is through the node server,
which is periodically polling the contest API for changes.

On the browser side there's currently no way to know when that backend data has
changed, so the scoreboard and clock are invalidating their data periodically to
force a refresh. This just adds the same auto-reloading to the teams/ and
problems/ pages. Everything should be reactive and auto refresh again.

Is there a better way? Almost definitely. :) The backend should probably be using
the event feed so that updates are immediate and there's no polling, and the
frontend should be using a websocket or even a periodic poll to see what or if
anything has changed instead of blindly reloading. But, that's for another day.

Also adds the missing height and overflow classes so that the problem page
scrolls correctly when there are lots of submissions.